// src/app/modules/admin/g-admin/bank/bank.types.ts
export interface BranchPagination {
    length: number;
    size: number;
    page: number;
    lastPage: number;
    startIndex: number;
    endIndex: number;
    pagination: {
        page: number,
        pageSize: number,
        pageCount: number,
        total: number
    }
}
// ...
export interface DataBank {
    code: number,
    status: boolean,
    message: string,
    data: {
        id:number,
        name: string,
        number: number,
        description: string,
        summary: number,
        bank_of: string,
        status: string,
        create_by: string,
        update_by: string,
        created_at: string,
        updated_at: string,
        bank_approves: [
            {
                id: number,
                bank_id: number,
                user_id: number,
                status: string,
                remark: string,
                create_by: string,
                update_by: string,
                created_at: string,
                updated_at: string
            },
        ],
    }
}